refactor: refactor the name of folders
