+++ src/(dashboard)/activity/activity-layout.tsx
